update to fftw-3.3.9

+++ src/fftw-3.3/api/mapflags.c
@@ -1,6 +1,6 @@
 /*
- * Copyright (c) 2003, 2007-11 Matteo Frigo
- * Copyright (c) 2003, 2007-11 Massachusetts Institute of Technology
+ * Copyright (c) 2003, 2007-14 Matteo Frigo
+ * Copyright (c) 2003, 2007-14 Massachusetts Institute of Technology
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -14,11 +14,11 @@
  *
  * You should have received a copy of the GNU General Public License
  * along with this program; if not, write to the Free Software
- * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  *
  */
 
-#include "api.h"
+#include "api/api.h"
 #include <math.h>
 
 /* a flag operation: x is either a flag, in which case xm == 0, or
@@ -43,9 +43,9 @@ typedef struct {
 #define NEQV(a, b) IMPLIES(YES(a), NO(b)), IMPLIES(NO(a), YES(b))
 
 static void map_flags(unsigned *iflags, unsigned *oflags,
-		      const flagop flagmap[], int nmap)
+		      const flagop flagmap[], size_t nmap)
 {
-     int i;
+     size_t i;
      for (i = 0; i < nmap; ++i)
           if (FLAGP(*iflags, flagmap[i].flag))
                *oflags = OP(*oflags, flagmap[i].op);
