+++ src/fftw-3.3/api/rdft2-pad.c
@@ -1,6 +1,6 @@
 /*
- * Copyright (c) 2003, 2007-11 Matteo Frigo
- * Copyright (c) 2003, 2007-11 Massachusetts Institute of Technology
+ * Copyright (c) 2003, 2007-14 Matteo Frigo
+ * Copyright (c) 2003, 2007-14 Massachusetts Institute of Technology
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -14,12 +14,12 @@
  *
  * You should have received a copy of the GNU General Public License
  * along with this program; if not, write to the Free Software
- * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  *
  */
 
 #include <string.h>
-#include "api.h"
+#include "api/api.h"
 
 const int *X(rdft2_pad)(int rnk, const int *n, const int *nembed,
 			int inplace, int cmplx, int **nfree)
@@ -28,8 +28,8 @@ const int *X(rdft2_pad)(int rnk, const int *n, const int *nembed,
      *nfree = 0;
      if (!nembed && rnk > 0) {
           if (inplace || cmplx) {
-               int *np = (int *) MALLOC(sizeof(int) * rnk, PROBLEMS);
-               memcpy(np, n, sizeof(int) * rnk);
+               int *np = (int *) MALLOC(sizeof(int) * (unsigned)rnk, PROBLEMS);
+               memcpy(np, n, sizeof(int) * (unsigned)rnk);
                np[rnk - 1] = (n[rnk - 1] / 2 + 1) * (1 + !cmplx);
                nembed = *nfree = np;
           } else
